pyproject.toml
[tool.poetry]
name = "backend"
version = "2.0.0"
description = "Charli3 Node Operator Backend."
authors = ["Deep Bhatt <deep.bhatt@charli3.io>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
pycardano = "0.10.0"
aiohttp = "^3.8.4"
python-json-logger = "^2.0.7"
pycryptodome = "^3.18.0"
charli3-offchain-core = {git = "https://***REMOVED***@github.com/Charli3-Official/charli3-offchain-core.git", rev = "v1.6.5"}
types-pyyaml = "^6.0.12.12"
sqlmodel = "^0.0.16"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
cachetools = "^5.3.3"

[tool.poetry.group.dev.dependencies]
mocket = "^3.11.0"
pytest = "^7.2.1"
boto3 = "^1.26.63"
pylint = "^2.16.1"
black = "^23.1.0"
pytest-asyncio = "^0.20.3"
sure = "^2.0.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
